removing of flush when cache is full
The job will be left to flush time activator, which will cause the coinview to be flushed within 3 minutes from consensus loop thread.

We need to remove the flush from here because it poses a risk of race condition, which was discussed in #426. It was partially eliminated in #498, but it seems that this code adds another possibility for new instance of #426.1.