Bugfix: occasional problem with node being not synced with its peers (#879)
* Improved logic of selecting if we should announce certain hash or not.

When deciding what headers to send we were checking if peer has header at the same height as one we want to send. If it does we would skip the header.
This was under the assumption that those headers are the same. A scenario when this is false was found so now we need to check if hashes are qual and if they don't we need to send the hash.

* Refactored foreach block. No logic was changed.

* Fixes based on the review.

* Code optimization.

* Some more optimizations.

* Code optimizations.

* Removed useless comment line.

* Added using FindAncestorOrSelf instead of GetAncestor

* Added using overload of FindAncestorOrSelf that already performs both height and hash checks.

* Fixed grammar.