Introduce an ErrorMessage property to TransactionBroadcastEntry.
If transaction fails during memory pool validation add error message to TransactionBroadcastEntry.

WalletManager to expose the TransactionBroadcastEntry (see GetTransactionBroadcastEntry()).  So it can bubble up to the client code via the WalletController.

Only process a valid transaction - in the WalletManager, if the State is not CantBroadCast.  Otherwise don't process and log the fact is failed.

Change Wallet integration tests based on above changes.