Support for json format to RPC GetBlock and included block height (#2497)
Added support for json format, included the annotations to make the fields the same as stratisX and bitcoin.  Therefore allowing the fullnode to be more easily swappable in to replace stratisX where connections made with rpc.

This is a behaviour change as it is now json format by default and optionally hex format.  This is because scenarios where the node is replacing a stratisX node, the json format is what is expected.