StratisX integration test capability (#2543)
We should probably have at least a few regression tests to ensure that SBFN remains compatible with stratisX for the forseeable future. This PR is a first step towards that.

Note that this PR does not include binaries for Linux or OSX yet (I can only realistically provide the Linux one myself). Ideally they should be rebuilt from source in a controlled environment and possibly signed, as unknown binaries are a security risk.

The stratisd.exe included here has a necessary fix by Maciej to allow operation on regtest. This would need to be merged into stratisX master.