[Light Wallet] Fix Breeze node and introduce pruning (#2966)
In this PR the node will now prune its database on startup and shut down.

An async loop will run that periodically deletes blocks from the block store. It will keep blocks up until 1000 blocks from the store tip.

On shutdown the database will be "compacted" so that all the deleted references are removed from the file and the database file re-saved.

On startup the node will check to see if its needs to be pruned (i.e. if the pruned tip is more than 1000 blocks from the store tip) if it is needed the database will be pruned and shrunk.

Some tests are still failing which I am looking at.

I have tested this extensively and the breeze wallet syncs fully.

The size of the DB after compactin is about 290kb.